Restructured project, added README
